remove legacy obstruction 2d env
